# src/app/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Documentação da API",
        "description": "Documentação do Projeto-Zeus, uma API para gerenciamento de membros, clientes, projetos e equipamentos.",
        "termsOfService": "http://swagger.io/terms/"
    },
    "host": "localhost:4001",
    "basePath": "/",
    "schemes": ["http"],
    "paths": {
...
        "/member": {
            "post": {
                "summary": "Criar membro",
                "parameters": [
                    { "name": "body", "in": "body", "required": true, "schema": { "type": "object", "properties": { "nomeCompleto": { "type": "string" }, "email": { "type": "string" }, "password": { "type": "string" }, "role": { "type": "string" }, "phone": { "type": "string" } }, "required": ["nomeCompleto", "email", "password", "role", "phone"] } }
                ],
                "responses": { "201": { "description": "Membro criado" }, "400": { "description": "Dados inválidos" } }
            },
            "get": {
                "summary": "Listar todos os membros",
                "responses": { "200": { "description": "Lista de membros" } }
            }
        },
...
        "/project": {
            "post": {
                "summary": "Criar projeto",
                "parameters": [
                    { "name": "body", "in": "body", "required": true, "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" }, "clientId": { "type": "integer" }, "memberId": { "type": "integer" } }, "required": ["name", "description", "clientId", "memberId"] } }
                ],
                "responses": { "201": { "description": "Projeto criado" }, "400": { "description": "Dados inválidos" }, "404": { "description": "Cliente não encontrado" } }
            },
            "get": {
                "summary": "Listar todos os projetos",
                "responses": { "200": { "description": "Lista de projetos" } }
            }
        },
...
        "/equipment": {
            "post": {
                "summary": "Criar equipamento",
                "parameters": [
                    { "name": "body", "in": "body", "required": true, "schema": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": "string" } }, "required": ["name", "description"] } }
                ],
                "responses": { "201": { "description": "Equipamento criado" }, "400": { "description": "Dados inválidos" } }
            },
            "get": {
                "summary": "Listar todos os equipamentos",
                "responses": { "200": { "description": "Lista de equipamentos" } }
            }
        },
...
        "/penalty": {
            "post": {
                "summary": "Criar penalidade",
                "parameters": [
                    { "name": "body", "in": "body", "required": true, "schema": { "type": "object", "properties": { "type": { "type": "string" }, "reason": { "type": "string" }, "date": { "type": "string", "format": "date" }, "memberId": { "type": "integer" } }, "required": ["type", "reason", "date", "memberId"] } }
                ],
                "responses": { "201": { "description": "Penalidade criada" }, "400": { "description": "Dados inválidos" } }
            },
            "get": {
                "summary": "Listar todas as penalidades",
                "responses": { "200": { "description": "Lista de penalidades" } }
            }
        },
...
    },
...
}